Animation Fix and utils creation
Fixed the janky animation mentioned on the previous branch merge.
Removed left and right of h1 and h2 and all elements fade in one after other. Table disappearing animation is fixed and after it disappears, diagnosis table fades in correctly.
Made a util file for the backend so now main flask app code is much cleaner to read and all of the functions of recommender system and predictor model are in utils file in backend.