Predictor Page built
Predictor Page or Neural Network page is finally built. It is similar to the other disease predictor page but with dynamic search instead of a predictor system. At this point everything is built before Chat page. It will have asynchronous load and everything.

Minor Improvements are still there to be made here.

The delete icon should be dynamic and should changed when hovered. The cursor should also change when hovered over that table's row.
simple feature to implement so do it fast.

Also changed the api just a bit. It now returns added and removed on every call with diagnosis.